requirements.txt
# Quant Strategy Research Engine
# All FREE libraries - no paid dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# WebSocket & HTTP
websockets>=12.0
httpx>=0.25.0

# Angel One SmartAPI SDK
smartapi-python>=1.4.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (SQLite - FREE)
aiosqlite>=0.19.0

# Numerical Computing
numpy>=1.24.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# TOTP for Angel One Auth
pyotp>=2.9.0

# HTML Templates
jinja2>=3.1.0

# Date/Time handling
python-dateutil>=2.8.0

# Logging (used by SmartAPI)
logzero>=1.7.0

# WebSocket client (used by SmartAPI)
websocket-client>=1.6.0

# Encryption (used by SmartAPI)
pycryptodome>=3.19.0
